refactor: 습관 달성률 계산 로직 helper함수로 분리
